Allow users to pass in redirect_url
I'm using DispatcherMiddleware to run multiple Dash applications within a single Flask app. Each of these Dash apps runs a separate Oauth context, and I need to be able to pass in a redirect url that is different for each of these apps. This PR enables users to optionally pass in a relative redirect URL.